use modal component and isolate display logic
